examples: selectively pull in config

+++ examples/glycine/minimization/run.sh
@@ -4,7 +4,8 @@ export GMX_MAXBACKUP=-1
 
 readonly NAME=gly_minimize
 
-cp ../gifs_config.ini .
+[ ! -f gifs_config.ini ] && cp ../gifs_config.ini .
+
 grompp -f minim.mdp -c ../glycine.gro -p ../glycine.top -n ../gly.ndx -o ${NAME}.tpr
 mdrun -v -deffnm ${NAME}
